Infinite World Generation
World generation was rewrote to be infinite. This is because optimizaitons later on require this and how we used to do calculations was only specific for smaller worlds. There is still a bug with some chunks not correctly hiding faces, but for now the world should be in theory infinite and working.